Fixed null object exception being thrown from Linq statement to retrieve AimSequenceNumber for the Earnings table.

+++ src/ESFA.DC.PeriodEnd.ReportService.Service/Builders/AppsMonthlyPaymentModelBuilder.cs
@@ -451,7 +451,7 @@ namespace ESFA.DC.PeriodEnd.ReportService.Service.Builders
                             if (paymentEarningEventId != null)
                             {
                                 // get the matching sequence number for this earning event id from the Earning Event table
-                                appsMonthlyPaymentModel.PaymentEarningEventAimSeqNumber = (byte)_appsMonthlyPaymentDasEarningsInfo?.Earnings
+                                appsMonthlyPaymentModel.PaymentEarningEventAimSeqNumber = _appsMonthlyPaymentDasEarningsInfo?.Earnings
                                    ?.SingleOrDefault(x => x?.EventId == paymentEarningEventId)?.LearningAimSequenceNumber;
                             }
                         }
